Move Treepath => Path and change compact string. (#164)
This PR contains several cleanups for Treepath.

Rename movetree.Treepath => movetree.Path. This is a bit more idiomatic
because this reduces stutter -- 'tree' is already in the package name.

Also, change the separators for CompactString representation of

. => -
: => x

Example: 1.2.0:3 => 1-2-0x3

I think this reads slightly better (0x3 is very clear, in my mind) and
also can be used directly in filenames so no FileSafeCompactString is
necessary.